requirements.txt
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.3
